automation performance tuning

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/dummy/ClientCommends.java
@@ -4,7 +4,7 @@ package org.apache.hadoop.hdfs.server.namenode.dummy;
 public class ClientCommends {
 	// Command: 0, finished. 1, received all MapRequest.
 	private int command = -1;
-	private int listSize = -1;
+	private long listSize = -1;
 	public int getCommand() {
 		return command;
 	}
@@ -13,11 +13,11 @@ public class ClientCommends {
 		this.command = command;
 	}
 
-  public int getListSize() {
+  public long getListSize() {
     return listSize;
   }
 
-  public void setListSize(int listSize) {
+  public void setListSize(long listSize) {
     this.listSize = listSize;
   }
